Extra ct the shared AI settings render helpers into an action-generic module
