+++ app/src/settings_view/ai_page.rs
@@ -18,7 +18,7 @@ use warp_core::ui::theme::color::internal_colors;
 use warp_editor::editor::NavigationKey;
 use warpui::elements::{
     Border, ChildAnchor, ChildView, ConstrainedBox, Container, CornerRadius, CrossAxisAlignment,
-    Dismiss, Empty, Expanded, Fill, Flex, FormattedTextElement, HighlightedHyperlink, Hoverable,
+    Dismiss, Empty, Expanded, Flex, FormattedTextElement, HighlightedHyperlink, Hoverable,
     HyperlinkLens, HyperlinkUrl, MainAxisAlignment, MainAxisSize, MouseStateHandle,
     OffsetPositioning, ParentAnchor, ParentElement, ParentOffsetBounds, Radius, Shrinkable, Stack,
     Text,
@@ -35,6 +35,10 @@ use warpui::{
     ViewHandle, WeakViewHandle, id,
 };
 
+use super::ai_shared::{
+    render_ai_feature_switch, render_ai_setting_description, render_ai_setting_label,
+    render_ai_setting_toggle, styles,
+};
 use super::custom_inference_modal::{
     CustomEndpointModal, CustomEndpointModalEvent, CustomEndpointModalViewState,
 };
@@ -44,9 +48,9 @@ use super::remove_custom_endpoint_confirmation_dialog::{
 };
 use super::set_default_model_modal::{SetDefaultModelModalBody, SetDefaultModelModalBodyEvent};
 use super::settings_page::{
-    HEADER_PADDING, InputListItem, LocalOnlyIconState, MatchData, PageType, SettingsPageMeta,
-    SettingsPageViewHandle, SettingsWidget, TOGGLE_BUTTON_RIGHT_PADDING, ToggleState,
-    build_sub_header, build_toggle_element, render_body_item_label,
+    CONTENT_FONT_SIZE, HEADER_PADDING, InputListItem, LocalOnlyIconState, MatchData, PageType,
+    SettingsPageMeta, SettingsPageViewHandle, SettingsWidget, TOGGLE_BUTTON_RIGHT_PADDING,
+    ToggleState, build_sub_header, build_toggle_element, render_body_item_label,
     render_body_item_label_with_icon, render_custom_size_header, render_dropdown_item,
     render_dropdown_item_label, render_filterable_dropdown_item, render_full_pane_width_ai_button,
     render_input_list, render_separator, render_settings_info_banner,
@@ -170,7 +174,6 @@ use crate::view_components::{Dropdown, DropdownItem};
 use crate::workspaces::workspace::{AdminEnablementSetting, CustomerType};
 use crate::{TelemetryEvent, UserWorkspaces, send_telemetry_from_ctx};
 
-const CONTENT_FONT_SIZE: f32 = 12.;
 const PRIMARY_HEADER_FONT_SIZE: f32 = 24.;
 
 const AI_SETTINGS_DROPDOWN_WIDTH: f32 = 250.;
@@ -4673,104 +4676,6 @@ impl From<ViewHandle<AISettingsPageView>> for SettingsPageViewHandle {
     }
 }
 
-fn render_ai_setting_toggle<S: Setting>(
-    label: impl Into<String>,
-    action: AISettingsPageAction,
-    is_setting_enabled: bool,
-    is_setting_toggleable: bool,
-    switch_state: SwitchStateHandle,
-    tooltip_states: &RefCell<HashMap<String, MouseStateHandle>>,
-    app: &AppContext,
-) -> Box<dyn Element> {
-    let appearance = Appearance::as_ref(app);
-    build_toggle_element(
-        render_body_item_label::<AISettingsPageAction>(
-            label.into(),
-            Some(styles::header_font_color(is_setting_toggleable, app)),
-            None,
-            LocalOnlyIconState::for_setting(
-                S::storage_key(),
-                S::sync_to_cloud(),
-                &mut tooltip_states.borrow_mut(),
-                app,
-            ),
-            ToggleState::Enabled,
-            appearance,
-        ),
-        render_ai_feature_switch(
-            switch_state,
-            is_setting_enabled,
-            is_setting_toggleable,
-            action,
-            app,
-        ),
-        appearance,
-        None,
-    )
-}
-
-fn render_ai_setting_label<S: Setting>(
-    label: impl Into<String>,
-    is_setting_toggleable: bool,
-    tooltip_states: &RefCell<HashMap<String, MouseStateHandle>>,
-    app: &AppContext,
-) -> Box<dyn Element> {
-    let appearance = Appearance::as_ref(app);
-    Container::new(render_body_item_label::<AISettingsPageAction>(
-        label.into(),
-        Some(styles::header_font_color(is_setting_toggleable, app)),
-        None,
-        LocalOnlyIconState::for_setting(
-            S::storage_key(),
-            S::sync_to_cloud(),
-            &mut tooltip_states.borrow_mut(),
-            app,
-        ),
-        ToggleState::Enabled,
-        appearance,
-    ))
-    .with_margin_bottom(HEADER_PADDING)
-    .finish()
-}
-
-fn render_ai_setting_description(
-    description: impl Into<Cow<'static, str>>,
-    is_setting_toggleable: bool,
-    app: &AppContext,
-) -> Box<dyn Element> {
-    let default_font_size = Appearance::as_ref(app).ui_font_size();
-    render_ai_setting_description_with_font_size(
-        description,
-        default_font_size,
-        is_setting_toggleable,
-        app,
-    )
-}
-
-fn render_ai_setting_description_with_font_size(
-    description: impl Into<Cow<'static, str>>,
-    font_size: f32,
-    is_setting_toggleable: bool,
-    app: &AppContext,
-) -> Box<dyn Element> {
-    let ui_builder = Appearance::as_ref(app).ui_builder();
-    ui_builder
-        .paragraph(description)
-        .with_style(UiComponentStyles {
-            font_size: Some(font_size),
-            font_color: Some(styles::description_font_color(is_setting_toggleable, app).into()),
-            margin: Some(
-                Coords::default()
-                    .top(styles::DESCRIPTION_NEGATIVE_MARGIN_OFFSET)
-                    .bottom(styles::DESCRIPTION_MARGIN_BOTTOM)
-                    .right(styles::TOGGLE_WIDTH_MARGIN),
-            ),
-            ..Default::default()
-        })
-        .build()
-        .finish()
-}
-
 fn render_toolbar_layout_editor(
     editor: &ViewHandle<AgentToolbarInlineEditor>,
     appearance: &Appearance,
@@ -4793,34 +4698,6 @@ fn render_toolbar_layout_editor(
         .finish();
 
     Flex::column().with_child(label).with_child(editor).finish()
-}
-
-fn render_ai_feature_switch(
-    state_handle: SwitchStateHandle,
-    is_setting_enabled: bool,
-    is_setting_toggleable: bool,
-    toggle_action: AISettingsPageAction,
-    app: &AppContext,
-) -> Box<dyn Element> {
-    let appearance = Appearance::as_ref(app);
-    let ui_builder = appearance.ui_builder();
-    ui_builder
-        .switch(state_handle)
-        .check(is_setting_enabled)
-        .with_disabled(!is_setting_toggleable)
-        .with_disabled_styles(UiComponentStyles {
-            background: Some(Fill::Solid(internal_colors::neutral_4(appearance.theme()))),
-            foreground: Some(Fill::Solid(internal_colors::neutral_5(appearance.theme()))),
-            ..Default::default()
-        })
-        .build()
-        .on_click(move |ctx, _, _| {
-            if !is_setting_toggleable {
-                return;
-            }
-            ctx.dispatch_typed_action(toggle_action.clone());
-        })
-        .finish()
 }
 
 fn render_ai_list(
@@ -10314,43 +10191,5 @@ impl SettingsWidget for CustomModelRoutersWidget {
         Container::new(column.finish())
             .with_margin_bottom(HEADER_PADDING)
             .finish()
-    }
-}
-
-mod styles {
-    use warp_core::ui::appearance::Appearance;
-    use warp_core::ui::theme::Fill;
-    use warpui::{AppContext, SingletonEntity};
-
-    // Apply a negative margin to the description text so it appears closer to the main
-    // settings option text.
-    pub const DESCRIPTION_NEGATIVE_MARGIN_OFFSET: f32 = -12.;
-
-    /// The space between a description and the next toggle.
-    pub const DESCRIPTION_MARGIN_BOTTOM: f32 = 12.;
-
-    /// Margin to leave for switch toggle to the right of the description subtext.
-    pub const TOGGLE_WIDTH_MARGIN: f32 = 48.;
-
-    pub fn header_font_color(is_enabled_setting: bool, app: &AppContext) -> Fill {
-        let appearance = Appearance::as_ref(app);
-        if is_enabled_setting {
-            appearance
-                .theme()
-                .main_text_color(appearance.theme().surface_2())
-        } else {
-            appearance.theme().disabled_ui_text_color()
-        }
-    }
-
-    pub fn description_font_color(is_enabled_setting: bool, app: &AppContext) -> Fill {
-        let appearance = Appearance::as_ref(app);
-        if is_enabled_setting {
-            appearance
-                .theme()
-                .sub_text_color(appearance.theme().surface_1())
-        } else {
-            appearance.theme().disabled_ui_text_color()
-        }
     }
 }
